COrrection diverses
 Modified Files:
 	glpi/update.php glpi/common/massiveaction.php
 	glpi/glpi/dicts/french.php glpi/ocsng/ocsng-import.php
 	glpi/ocsng/ocsng-link.php glpi/ocsng/ocsng-sync.php
 	glpi/setup/setup-dropdowns.php


git-svn-id: https://forge.indepnet.net/svn/glpi/trunk@2748 35612d4c-9d10-0410-837c-bf9dd1be60d7

+++ update.php
@@ -3691,12 +3691,13 @@ if(!FieldExists("glpi_licenses","comments")) {
 
 ///////////// MODE OCS
 
-if(TableExists("glpi_ocs_link")) {
+// Delete plugin table
+if(TableExists("glpi_ocs_link")&&!FieldExists("glpi_ocs_link","import_device")) {
 	$query = "DROP TABLE `glpi_ocs_link`";
 	$db->query($query) or die("0.65 MODE OCS drop plugin ocs_link ".$lang["update"][90].$db->error());
 }
 
-if(TableExists("glpi_ocs_config")) {
+if(TableExists("glpi_ocs_config")&&!FieldExists("glpi_ocs_config","checksum")) {
 	$query = "DROP TABLE `glpi_ocs_config`";
 	$db->query($query) or die("0.65 MODE OCS drop plugin ocs_config ".$lang["update"][90].$db->error());
 }
